requirements.txt
autopep8==1.5.7
azure-core==1.15.0
azure-cosmos==4.2.0
azure-functions==1.7.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
